Made the test compilable with clang on Windows

+++ engine/testing/logChessStructs.c
@@ -212,7 +212,7 @@ void printPosition(ChessPosition position, int score, Move bestMove, float stock
 
 void printBitBoard(u64 bitboard) {
     for (int i = 0; i < 64; i++) {
-        printf("%ld", (bitboard >> i) & 1);
+        printf("%d", (int) ((bitboard >> i) & 1));
         if ((i + 1) % 8 == 0) {
             printf("\n");
         }
@@ -222,7 +222,7 @@ void printBitBoard(u64 bitboard) {
 void printBin(const u64 num) {
     printf("0b");
     for (int i = 63; i >= 0; i--) {
-        printf("%lu", (num >> i) & 1);
+        printf("%d", (int) ((num >> i) & 1));
     }
     printf("\n");
 }
